feat(mobile): export notes as PDF via share sheet
Adds an 'Export as PDF' action to the note menu, mirroring desktop PDF
export — a beautifully formatted PDF you can print, save to Files, or
share.

- shared/notes/pdf-template.ts: the desktop PDF HTML template moved to
  shared (pure, no Electron deps); electron/lib/pdf-template.ts is now a
  re-export shim so both platforms build the same document.
- shared/notes/note-html.ts: pure helpers converting Cairn callouts
  (> [!type]) and [[wikilinks]] into the exact [data-callout] / chip HTML
  the template styles — so mobile PDFs match desktop, incl. callouts +
  wikilinks. Unit-tested.
- mobile/src/notes/note-pdf.ts: renders note markdown -> HTML (markdown-it
  + markdown-it-mark + the shared callout/wikilink passes) -> buildPdfHtml
  -> expo-print printToFileAsync -> share sheet (follows light/dark).
- Wired an Export as PDF menu action (ICON_PDF) into NoteDetailScreen.
- deps: added expo-print + expo-sharing (SDK-57 pinned); bumped safe
  JS-only libs (ai, lucide-react-native). NOTE: removed the expo-sharing
  config plugin from app.json — its plugin scaffolds an unwanted iOS share
  extension; shareAsync needs no plugin (this caused a launch crash).

Verify: expo-doctor clean, mobile tsc + eslint clean, shared tests green,
desktop compile OK. Changelog mobile v0.1.3.
