+++ mobile/changelogs/v0.1.3.md
@@ -4,6 +4,7 @@ Polish for the workspace switcher and Knowledge Graph controls.
 
 ### New
 
+- **Export a note as a PDF**: The note menu now has an **Export as PDF** action alongside Export as Markdown. It renders the note into a beautifully formatted PDF — the same layout, typography, code blocks, tables, and callouts as the desktop PDF export (following your light/dark appearance) — then opens the system share sheet so you can Print it, save it to Files, or send it via Messages/Mail/AirDrop. Callouts (`> [!note]`, `[!warning]`, etc.) and `[[wikilinks]]` render with the same styling as desktop.
 - **Long-press a note to move, pin, or delete it**: Press and hold any note in a project's Notes list to open a quick actions menu — **Pin/Unpin**, **Move to project**, **Move to folder**, or **Delete** — mirroring the note menu on desktop. "Move to project" and "Move to folder" open a picker of your other projects / this project's folders. This is the first way to move a note between projects on mobile (there was previously no manual way to do it at all).
 - **Project Overview on mobile**: Opening a project now starts on a new **Overview** tab (alongside Notes and Board) that gives you an at-a-glance snapshot — a completion ring, note / open-task / overdue counts, a breakdown of tasks by column and by priority, what's due soon, your pinned and recent notes, and a recent-activity feed. Tapping anything jumps straight to it. This mirrors the desktop project Overview.
 - **Drag to reschedule on the calendar**: Long-press a task on the calendar and drag it onto another day to change its due date — the same lift-and-drop gesture as the board. There's also a new **Unscheduled** tray at the bottom listing tasks with no due date: drag one onto a day to schedule it, or drag a dated task into the tray to clear its due date. Available on both the workspace and per-project calendars.
