+++ mobile/package.json
@@ -18,29 +18,31 @@
   },
   "dependencies": {
     "@expo/material-symbols": "^0.1.1",
-    "@expo/ui": "~57.0.3",
+    "@expo/ui": "~57.0.6",
     "@react-native-community/datetimepicker": "9.1.0",
-    "ai": "^7.0.15",
-    "expo": "~57.0.2",
-    "expo-build-properties": "~57.0.3",
-    "expo-clipboard": "~57.0.0",
+    "ai": "^7.0.29",
+    "expo": "~57.0.6",
+    "expo-build-properties": "~57.0.5",
+    "expo-clipboard": "~57.0.1",
     "expo-constants": "~57.0.2",
-    "expo-document-picker": "~57.0.0",
+    "expo-document-picker": "~57.0.1",
     "expo-file-system": "~57.0.0",
-    "expo-glass-effect": "~57.0.0",
-    "expo-haptics": "~57.0.0",
-    "expo-image-picker": "~57.0.2",
-    "expo-linking": "~57.0.1",
-    "expo-router": "~57.0.3",
-    "expo-secure-store": "~57.0.0",
-    "expo-splash-screen": "~57.0.2",
-    "expo-sqlite": "~57.0.0",
-    "expo-status-bar": "~57.0.0",
-    "expo-updates": "~57.0.6",
+    "expo-glass-effect": "~57.0.1",
+    "expo-haptics": "~57.0.1",
+    "expo-image-picker": "~57.0.4",
+    "expo-linking": "~57.0.3",
+    "expo-print": "~57.0.1",
+    "expo-router": "~57.0.6",
+    "expo-secure-store": "~57.0.1",
+    "expo-sharing": "~57.0.5",
+    "expo-splash-screen": "~57.0.4",
+    "expo-sqlite": "~57.0.1",
+    "expo-status-bar": "~57.0.1",
+    "expo-updates": "~57.0.7",
     "js-tiktoken": "^1.0.21",
     "katex": "^0.17.0",
     "lowlight": "^3.3.0",
-    "lucide-react-native": "^1.23.0",
+    "lucide-react-native": "^1.24.0",
     "markdown-it-mark": "^4.0.0",
     "mermaid": "^11.16.0",
     "react": "19.2.3",
@@ -50,7 +52,7 @@
     "react-native-gesture-handler": "~2.32.0",
     "react-native-keyboard-controller": "1.21.9",
     "react-native-markdown-display": "^7.0.2",
-    "react-native-reanimated": "~4.5.0",
+    "react-native-reanimated": "4.5.0",
     "react-native-safe-area-context": "~5.7.0",
     "react-native-screens": "4.25.2",
     "react-native-svg": "15.15.4",
@@ -62,7 +64,7 @@
     "@types/react": "~19.2.4",
     "babel-plugin-module-resolver": "^5.0.2",
     "d3": "^7.9.0",
-    "eslint": "^9.39.4",
+    "eslint": "^9.39.5",
     "eslint-config-expo": "~57.0.0",
     "typescript": "~6.0.3"
   },
