+++ mobile/src/components/toolbar-icons.ts
@@ -28,6 +28,7 @@ import PinXml from "@expo/material-symbols/keep.xml";
 import UnpinXml from "@expo/material-symbols/keep_off.xml";
 import TagXml from "@expo/material-symbols/sell.xml";
 import ShareXml from "@expo/material-symbols/ios_share.xml";
+import PdfXml from "@expo/material-symbols/picture_as_pdf.xml";
 import CalendarXml from "@expo/material-symbols/calendar_month.xml";
 import CalendarViewMonthXml from "@expo/material-symbols/calendar_view_month.xml";
 import CalendarViewWeekXml from "@expo/material-symbols/calendar_view_week.xml";
@@ -70,6 +71,8 @@ export const ICON_SETTINGS: ToolbarIcon = ios ? "gearshape" : SettingsXml;
 export const ICON_AI: ToolbarIcon = ios ? "brain" : NeurologyXml;
 /** Export / share (share sheet). */
 export const ICON_EXPORT: ToolbarIcon = ios ? "square.and.arrow.up" : ShareXml;
+/** Export as PDF. */
+export const ICON_PDF: ToolbarIcon = ios ? "doc.richtext" : PdfXml;
 /** iCloud sync — synced/idle. */
 export const ICON_ICLOUD: ToolbarIcon = ios ? "icloud" : CloudXml;
 /** iCloud sync — offline / unavailable. */
